+++ modules/aws/eks-global/variables.tf
@@ -19,39 +19,6 @@ variable "unique_suffix" {
   type        = string
 }
 
-variable "namespaces" {
-  description = "The namespaces that should be created in Kubernetes"
-  type = list(
-    object({
-      name                    = string
-      delegate_resource_group = bool
-    })
-  )
-}
-
-variable "azure_ad_group_prefix" {
-  description = "Prefix for Azure AD Groups"
-  type        = string
-  default     = "az"
-}
-
-variable "group_name_separator" {
-  description = "Separator for group names"
-  type        = string
-  default     = "-"
-}
-
-variable "subscription_name" {
-  description = "The commonName for the subscription"
-  type        = string
-}
-
-variable "eks_group_name_prefix" {
-  description = "Prefix for EKS Azure AD groups"
-  type        = string
-  default     = "eks"
-}
-
 variable "eks_cloudwatch_retention_period" {
   description = "eks cloudwatch retention period"
   type        = number