+++ modules/kubernetes/aks-core/main.tf
@@ -53,3 +53,7 @@ data "azurerm_client_config" "current" {}
 data "azurerm_resource_group" "this" {
   name = "rg-${var.environment}-${var.location_short}-${var.name}"
 }
+
+data "azurerm_resource_group" "global" {
+  name = "rg-${var.environment}-${var.global_location_short}-global"
+}