+++ validation/azure/aks/main.tf
@@ -21,6 +21,10 @@ module "aks" {
   core_name                     = "core"
   aks_name_suffix               = "1"
   aks_managed_identity_group_id = "1337"
+  azure_metrics_identity = {
+    id           = "id"
+    principal_id = "pid"
+  }
 
   aks_config = {
     version          = "1.21.9"